+++ experimental/packages/opentelemetry-sdk-node/src/index.ts
@@ -23,3 +23,4 @@ export { NodeSDK } from './sdk';
 export type { LoggerProviderConfig, MeterProviderConfig } from './sdk';
 export type { NodeSDKConfiguration } from './types';
 export { startNodeSDK } from './start';
+export { buildSamplerFromConfig } from './utils';
