+++ packages/opentelemetry-sdk-trace-base/package.json
@@ -90,7 +90,6 @@
     "@opentelemetry/api": ">=1.3.0 <1.10.0"
   },
   "dependencies": {
-    "@opentelemetry/configuration": "0.214.0",
     "@opentelemetry/core": "2.6.1",
     "@opentelemetry/resources": "2.6.1",
     "@opentelemetry/semantic-conventions": "^1.29.0"
