+++ packages/opentelemetry-sdk-trace-base/src/index.ts
@@ -4,7 +4,6 @@
  */
 
 export { BasicTracerProvider } from './BasicTracerProvider';
-export { buildSamplerFromConfig } from './config';
 export { BatchSpanProcessor, RandomIdGenerator } from './platform';
 export { ConsoleSpanExporter } from './export/ConsoleSpanExporter';
 export { InMemorySpanExporter } from './export/InMemorySpanExporter';
