+++ packages/opentelemetry-sdk-trace-base/tsconfig.json
@@ -19,9 +19,6 @@
       "path": "../../api"
     },
     {
-      "path": "../../experimental/packages/configuration"
-    },
-    {
       "path": "../../semantic-conventions"
     },
     {
